[Feat] BaseTextFieldState를 통해서 텍스트필드 상태관리 구현 #87
