+++ ZOOC/ZOOC/Global/Extension/UITextField+.swift
@@ -19,7 +19,6 @@ extension UITextField {
         let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: inset, height: self.frame.height))
         self.rightView = paddingView
         self.rightViewMode = ViewMode.always
-        
     }
     
 }
